Correct branch length calculation
The proposal distribution specifies that a height
increment should be sampled from an exponential
distribution and that the branch lengths are
computed based on the height increment. The
current solution was incorrectly sampling the
branch lengths from an exponential distribution
directly.

This required replacing the tree node
'pll_rnode_s' with 'phylo_tree_node' to keep track
of the tree height at any node.